Refactor: Break down Tilemap.cs into focused modular components
- Split large Tilemap.cs (1194 lines) into 7 focused modules
- Created TilemapCollection.cs for multi-map management
- Created AnimatedTile.cs for complete animation system
- Created TileLayer.cs and ObjectLayer.cs for data structures
- Created TilesetDefinition.cs for tileset metadata
- Created Tilemap.JsonLoader.cs as partial class for JSON parsing
- Reduced core Tilemap.cs to 405 lines with only rendering/access logic
- Updated README.md to document new modular architecture
- Maintains all existing public APIs - no breaking changes
- Improves maintainability through single responsibility principle
- Build verified - all modules compile successfully
